Rename EditCard to FurnitureEditCard
