+++ src/components/FurnitureEditCard/AttributeQuestions.vue
