+++ src/components/FurnitureEditCard/PhysicalAttributes.vue
